+++ authress_sdk/models/access_record_user.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-class AccessRecordUsers(object):
+class AccessRecordUser(object):
 
     """
     Attributes:
@@ -16,26 +16,26 @@ class AccessRecordUsers(object):
     }
 
     def __init__(self, user_id=None):
-        """AccessRecordUsers"""
+        """AccessRecordUser"""
         self._user_id = None
         self.user_id = user_id
 
     @property
     def user_id(self):
-        """Gets the user_id of this AccessRecordUsers.
+        """Gets the user_id of this AccessRecordUser.
 
 
-        :return: The user_id of this AccessRecordUsers.
+        :return: The user_id of this AccessRecordUser.
         :rtype: str
         """
         return self._user_id
 
     @user_id.setter
     def user_id(self, user_id):
-        """Sets the user_id of this AccessRecordUsers.
+        """Sets the user_id of this AccessRecordUser.
 
 
-        :param user_id: The user_id of this AccessRecordUsers.
+        :param user_id: The user_id of this AccessRecordUser.
         :type: str
         """
         if user_id is None: